refactor(ses): extract SesContactService from SesService (strangler step 5)
Fifth store-based domain split, the first multi-store one:
one @ApplicationScoped service owns both the contact-list and contact stores
plus their locks, collapsing two SesService constructor arguments into one.
It also owns the list-management contact behaviour (auto-create-on-send,
opt-out evaluation, unsubscribe); the send orchestration
(collectListManagementOptOuts) stays in the facade and calls into the
service, keeping the shared extractEmailAddress helper there. SesController's
two ContactWithList/ContactsWithList type references move to the service;
otherwise consumers are unchanged.
